symbols: major rework of how intrinsics are sent
Swapped out raw interned symbols for property symbols to prevent lisp's
printer from accidentally prepending the package name, which was causing
all sorts of problems when any of these functions were called from
outside of the cl-gypsum-client package.

This necessitated a slight design change to the CommandParser.java file
in the gypsum server.

This is all pending a rewrite to embed a proper lisp evaluator on the
android side and to reverse the connection direction (servers are
actually clients, and should connect to the silica server on the motes).